Update S104: "Too many lines in a file" should only count lines of code
