+++ sonaranalyzer-dotnet/src/SonarAnalyzer.CSharp/Rules/BooleanLiteralUnnecessaryCodeFixProvider.cs
@@ -189,16 +189,16 @@ namespace SonarAnalyzer.Rules.CSharp
             #region logical and false, logical or true
 
             if (binary.IsKind(SyntaxKind.LogicalAndExpression) &&
-                (EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.FalseLiteralExpression) ||
-                 EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.FalseLiteralExpression)))
+                (EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.FalseLiteralExpression) ||
+                 EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.FalseLiteralExpression)))
             {
-                return SyntaxHelper.FalseLiteralExpression;
+                return CSharpSyntaxHelper.FalseLiteralExpression;
             }
             if (binary.IsKind(SyntaxKind.LogicalOrExpression) &&
-                (EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.TrueLiteralExpression) ||
-                 EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.TrueLiteralExpression)))
+                (EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.TrueLiteralExpression) ||
+                 EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.TrueLiteralExpression)))
             {
-                return SyntaxHelper.TrueLiteralExpression;
+                return CSharpSyntaxHelper.TrueLiteralExpression;
             }
 
             #endregion
@@ -208,28 +208,28 @@ namespace SonarAnalyzer.Rules.CSharp
             if (binary.IsKind(SyntaxKind.EqualsExpression) &&
                 TwoSidesAreDifferentBooleans(binary))
             {
-                return SyntaxHelper.FalseLiteralExpression;
+                return CSharpSyntaxHelper.FalseLiteralExpression;
             }
             if (binary.IsKind(SyntaxKind.EqualsExpression) &&
                 TwoSidesAreSameBooleans(binary))
             {
-                return SyntaxHelper.TrueLiteralExpression;
+                return CSharpSyntaxHelper.TrueLiteralExpression;
             }
             if (binary.IsKind(SyntaxKind.NotEqualsExpression) &&
                 TwoSidesAreSameBooleans(binary))
             {
-                return SyntaxHelper.FalseLiteralExpression;
+                return CSharpSyntaxHelper.FalseLiteralExpression;
             }
             if (binary.IsKind(SyntaxKind.NotEqualsExpression) &&
                 TwoSidesAreDifferentBooleans(binary))
             {
-                return SyntaxHelper.TrueLiteralExpression;
+                return CSharpSyntaxHelper.TrueLiteralExpression;
             }
 
             #endregion
 
-            if (EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.TrueLiteralExpression) ||
-                EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.FalseLiteralExpression))
+            if (EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.TrueLiteralExpression) ||
+                EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.FalseLiteralExpression))
             {
                 return binary.Right;
             }
@@ -239,33 +239,33 @@ namespace SonarAnalyzer.Rules.CSharp
         private static bool TwoSidesAreDifferentBooleans(BinaryExpressionSyntax binary)
         {
             return (
-                EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.TrueLiteralExpression) &&
-                EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.FalseLiteralExpression)) ||
+                EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.TrueLiteralExpression) &&
+                EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.FalseLiteralExpression)) ||
                 (
-                EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.FalseLiteralExpression) &&
-                EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.TrueLiteralExpression));
+                EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.FalseLiteralExpression) &&
+                EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.TrueLiteralExpression));
         }
         private static bool TwoSidesAreSameBooleans(BinaryExpressionSyntax binary)
         {
             return (
-                EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.TrueLiteralExpression) &&
-                EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.TrueLiteralExpression)) ||
+                EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.TrueLiteralExpression) &&
+                EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.TrueLiteralExpression)) ||
                 (
-                EquivalenceChecker.AreEquivalent(binary.Left, SyntaxHelper.FalseLiteralExpression) &&
-                EquivalenceChecker.AreEquivalent(binary.Right, SyntaxHelper.FalseLiteralExpression));
+                EquivalenceChecker.AreEquivalent(binary.Left, CSharpSyntaxHelper.FalseLiteralExpression) &&
+                EquivalenceChecker.AreEquivalent(binary.Right, CSharpSyntaxHelper.FalseLiteralExpression));
         }
 
         private static Document RemovePrefixUnary(Document document, SyntaxNode root,
             SyntaxNode literal)
         {
-            if (EquivalenceChecker.AreEquivalent(literal, SyntaxHelper.TrueLiteralExpression))
+            if (EquivalenceChecker.AreEquivalent(literal, CSharpSyntaxHelper.TrueLiteralExpression))
             {
-                var newRoot = root.ReplaceNode(literal.Parent, SyntaxHelper.FalseLiteralExpression);
+                var newRoot = root.ReplaceNode(literal.Parent, CSharpSyntaxHelper.FalseLiteralExpression);
                 return document.WithSyntaxRoot(newRoot);
             }
             else
             {
-                var newRoot = root.ReplaceNode(literal.Parent, SyntaxHelper.TrueLiteralExpression);
+                var newRoot = root.ReplaceNode(literal.Parent, CSharpSyntaxHelper.TrueLiteralExpression);
                 return document.WithSyntaxRoot(newRoot);
             }
         }
@@ -273,7 +273,7 @@ namespace SonarAnalyzer.Rules.CSharp
         private static Document RemoveConditional(Document document, SyntaxNode root,
             ConditionalExpressionSyntax conditional)
         {
-            if (EquivalenceChecker.AreEquivalent(conditional.WhenTrue, SyntaxHelper.TrueLiteralExpression))
+            if (EquivalenceChecker.AreEquivalent(conditional.WhenTrue, CSharpSyntaxHelper.TrueLiteralExpression))
             {
                 var newRoot = root.ReplaceNode(conditional,
                     conditional.Condition.WithAdditionalAnnotations(Formatter.Annotation));
@@ -302,7 +302,7 @@ namespace SonarAnalyzer.Rules.CSharp
         {
             var whenTrue = conditional.WhenTrue.RemoveParentheses();
             if (whenTrue.Equals(syntaxNode) &&
-                EquivalenceChecker.AreEquivalent(syntaxNode, SyntaxHelper.TrueLiteralExpression))
+                EquivalenceChecker.AreEquivalent(syntaxNode, CSharpSyntaxHelper.TrueLiteralExpression))
             {
                 var newRoot = ReplaceExpressionWithBinary(conditional, root,
                     SyntaxKind.LogicalOrExpression,
@@ -313,7 +313,7 @@ namespace SonarAnalyzer.Rules.CSharp
             }
 
             if (whenTrue.Equals(syntaxNode) &&
-                EquivalenceChecker.AreEquivalent(syntaxNode, SyntaxHelper.FalseLiteralExpression))
+                EquivalenceChecker.AreEquivalent(syntaxNode, CSharpSyntaxHelper.FalseLiteralExpression))
             {
                 var newRoot = ReplaceExpressionWithBinary(conditional, root,
                     SyntaxKind.LogicalAndExpression,
@@ -326,7 +326,7 @@ namespace SonarAnalyzer.Rules.CSharp
             var whenFalse = conditional.WhenFalse.RemoveParentheses();
 
             if (whenFalse.Equals(syntaxNode) &&
-                EquivalenceChecker.AreEquivalent(syntaxNode, SyntaxHelper.TrueLiteralExpression))
+                EquivalenceChecker.AreEquivalent(syntaxNode, CSharpSyntaxHelper.TrueLiteralExpression))
             {
                 var newRoot = ReplaceExpressionWithBinary(conditional, root,
                     SyntaxKind.LogicalOrExpression,
@@ -337,7 +337,7 @@ namespace SonarAnalyzer.Rules.CSharp
             }
 
             if (whenFalse.Equals(syntaxNode) &&
-                EquivalenceChecker.AreEquivalent(syntaxNode, SyntaxHelper.FalseLiteralExpression))
+                EquivalenceChecker.AreEquivalent(syntaxNode, CSharpSyntaxHelper.FalseLiteralExpression))
             {
                 var newRoot = ReplaceExpressionWithBinary(conditional, root,
                     SyntaxKind.LogicalAndExpression,
