+++ sonaranalyzer-dotnet/src/SonarAnalyzer.CSharp/Rules/GetTypeWithIsAssignableFrom.cs
@@ -113,7 +113,7 @@ namespace SonarAnalyzer.Rules.CSharp
         private static void CheckAsOperatorComparedToNull(ExpressionSyntax sideA, ExpressionSyntax sideB, Location location,
             SyntaxNodeAnalysisContext context)
         {
-            if (!EquivalenceChecker.AreEquivalent(sideA.RemoveParentheses(), SyntaxHelper.NullLiteralExpression))
+            if (!EquivalenceChecker.AreEquivalent(sideA.RemoveParentheses(), CSharpSyntaxHelper.NullLiteralExpression))
             {
                 return;
             }
