+++ sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/ResourceTests/GeneratedResourcesTest.cs
@@ -1,22 +1,23 @@
-﻿/*
- * SonarAnalyzer for .NET
- * Copyright (C) 2015-2017 SonarSource SA
- * mailto: contact AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
+﻿extern alias vbnet;
+/*
+* SonarAnalyzer for .NET
+* Copyright (C) 2015-2017 SonarSource SA
+* mailto: contact AT sonarsource DOT com
+*
+* This program is free software; you can redistribute it and/or
+* modify it under the terms of the GNU Lesser General Public
+* License as published by the Free Software Foundation; either
+* version 3 of the License, or (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+* Lesser General Public License for more details.
+*
+* You should have received a copy of the GNU Lesser General Public License
+* along with this program; if not, write to the Free Software Foundation,
+* Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+*/
 
 using System;
 using System.Collections.Generic;
@@ -29,6 +30,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarAnalyzer.Common;
 using SonarAnalyzer.Helpers;
 using SonarAnalyzer.Rules;
+using vbnet::SonarAnalyzer.Helpers.VisualBasic;
 
 namespace SonarAnalyzer.UnitTest.ResourceTests
 {
@@ -40,7 +42,7 @@ namespace SonarAnalyzer.UnitTest.ResourceTests
         {
             var rulesFromResources = GetRulesFromResources(@"..\..\..\..\..\..\rspec\cs");
 
-            var rulesFromClasses = GetRulesFromClasses(typeof(SyntaxHelper).Assembly);
+            var rulesFromClasses = GetRulesFromClasses(typeof(CSharpSyntaxHelper).Assembly);
 
             rulesFromResources.Should().Equal(rulesFromClasses);
         }
@@ -50,7 +52,7 @@ namespace SonarAnalyzer.UnitTest.ResourceTests
         {
             var rulesFromResources = GetRulesFromResources(@"..\..\..\..\..\..\rspec\vbnet");
 
-            var rulesFromClasses = GetRulesFromClasses(typeof(SonarAnalyzer.Helpers.VisualBasic.SyntaxHelper).Assembly);
+            var rulesFromClasses = GetRulesFromClasses(typeof(VisualBasicSyntaxHelper).Assembly);
 
             rulesFromResources.Should().Equal(rulesFromClasses);
         }
