--- sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/TestCases/FileLines12.cs
@@ -1,11 +1,0 @@
-﻿namespace Tests.Diagnostics
-{
-    public class FileLines12
-    {
-        public FileLines12()
-        {
-        }
-    }
-}
-
-// hello
