--- sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/TestCases/FileLines12.vb
@@ -1,11 +1,0 @@
-﻿Namespace Ns
-    Class PlaceholderClass
-        Sub New()
-
-        End Sub
-    End Class
-
-End Namespace
-
-
-' Some comment here
