--- sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/TestCases/FileLines13.cs
@@ -1,12 +1,0 @@
-﻿// Noncompliant
-namespace Tests.Diagnostics
-{
-    public class FileLines
-    {
-        public FileLines()
-        {
-        }
-    }
-}
-
-// hello
