--- sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/TestCases/FileLines13.vb
@@ -1,12 +1,0 @@
-﻿' Noncompliant
-Namespace Ns
-    Class PlaceholderClass
-        Sub New()
-
-        End Sub
-    End Class
-
-End Namespace
-
-
-' Some comment here
