+++ sonaranalyzer-dotnet/src/Tests/SonarAnalyzer.UnitTest/TestCases/FileLines20.cs
@@ -1,0 +1,29 @@
+﻿// Noncompliant {{This file has 20 lines, which is greater than 10 authorized. Split it into smaller files.}}
+namespace Tests.Diagnostics
+{
+    public class FooBar
+    {
+        public FooBar()
+        {
+            System
+                   .
+                   Collections
+
+                   .
+                   // WTF
+                   Generic
+
+                   .
+                   List<int>
+                   .
+                   // Who writes this kind of code!
+                   Enumerator
+
+                   .
+
+                   Equals(null);
+        }
+    }
+}
+
+// hello
